Updated Spec to include Elicitation. Code refactoring included.

+++ src/main/java/za/co/sindi/ai/mcp/mapper/JsonProgressTokenSerialization.java
@@ -43,7 +43,7 @@ public class JsonProgressTokenSerialization implements JsonbSerializer<ProgressT
 		if (progressToken == null) {
 			try {
 				if (parser.isIntegralNumber()) {
-					progressToken = ProgressToken.of(parser.getLong());
+					progressToken = ProgressToken.of(parser.getInt());
 				}
 			} catch (IllegalStateException e) {
 				progressToken = ProgressToken.of(parser.getString());
